requirements.txt
web3
requests
python-dotenv
numpy
pandas
openai
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
psutil